Adjustments and New Features
All NPCs who the player can have contact has its own colliders. When the player runs into a NPC, or the other way around, the vest vibrates.

The colliders were decreased in every side (height, width and depth) for the proximity and more realistic interaction.

Now, the player starts on a bench and with a NPC on his/her side.

The working of colliders were reconfigured to use OnTriggerEnter instead of OnTriggerStay to not drop the FPS of the game when a collision happens.

All the station walls that the player can run into are collidable and can be felt with the vest. However, when the colliders enter the wall, the vest stops vibrating because the colliders call the stop vest function, so the vest stops and this issue couldn't be worked around yet.