Data management update
The name is taken from the console.

It is no longer saving the date because it is possible to get it from the startTime variable.

It saves the start and end times in a Date format.

There are two new boolean variables to save whether the user's heart rate was high or low during the experience.

With the heart rate, it saves the time of each heartbeat.

When the watch is disconnected, it only saves the last date. If the application is forced to close, it checks for a final time and saves the current time if none exists.